refactor workbench risk concentration route
