+++ src/app/router_registry.py
@@ -102,6 +102,9 @@ from app.routers.workbench_performance_modules import (
     router as workbench_performance_modules_router,
 )
 from app.routers.workbench_risk import router as workbench_risk_router
+from app.routers.workbench_risk_concentration import (
+    router as workbench_risk_concentration_router,
+)
 from app.routers.workbench_sandbox import router as workbench_sandbox_router
 
 
@@ -157,6 +160,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(workbench_performance_evidence_router)
     app.include_router(workbench_performance_modules_router)
     app.include_router(workbench_risk_router)
+    app.include_router(workbench_risk_concentration_router)
     app.include_router(workbench_sandbox_router)
     app.include_router(reporting_router)
     app.include_router(reporting_snapshots_router)
