+++ src/app/routers/workbench_risk.py
@@ -4,22 +4,16 @@ from fastapi import APIRouter, Header, Path, Query
 
 from app.contracts.risk_workspace import (
     WorkbenchRiskAttributionResponse,
-    WorkbenchRiskConcentrationResponse,
     WorkbenchRiskDrawdownResponse,
     WorkbenchRiskRollingResponse,
     WorkbenchRiskSummaryResponse,
 )
 from app.middleware.correlation import correlation_id_var
+from app.routers.workbench_risk_common import RISK_PERIOD_QUERY_DESCRIPTION
 from app.services.caller_context import caller_context_headers
 from app.services.workbench_service_provider import risk_workspace_service
 
 router = APIRouter(prefix="/api/v1/workbench", tags=["workbench"])
-
-RISK_PERIOD_QUERY_DESCRIPTION = (
-    "Canonical risk horizon. Use platform-governed values such as MTD, QTD, YTD, 1Y, 3Y, 5Y, "
-    "SI, YEAR, or EXPLICIT. Legacy aliases ONE_YEAR, THREE_YEAR, FIVE_YEAR, and ITD may be "
-    "accepted for compatibility but are normalized before calling lotus-risk."
-)
 
 
 def _required_caller_context(
@@ -124,73 +118,6 @@ async def get_workbench_risk_summary(
         report_start_date=report_start_date,
         report_end_date=report_end_date,
         reporting_currency=reporting_currency,
-    )
-
-
-@router.get(
-    "/{portfolio_id}/risk/concentration",
-    response_model=WorkbenchRiskConcentrationResponse,
-    summary="Get Workbench Risk Concentration",
-    description=(
-        "Returns Gateway-shaped, stateful lotus-risk concentration analytics for Workbench "
-        "position, issuer, and coverage concentration review. Use this route when the user "
-        "needs issuer mapping coverage, top-position concentration, or concentration posture "
-        "beyond the headline risk summary. Simulation concentration remains gated to a future "
-        "sandbox-aware slice."
-    ),
-)
-async def get_workbench_risk_concentration(
-    portfolio_id: str = Path(
-        ...,
-        description=(
-            "Canonical portfolio identifier for the stateful workbench risk concentration surface."
-        ),
-        examples=["PF_1001"],
-    ),
-    period: str = Query(
-        default="YTD",
-        description=RISK_PERIOD_QUERY_DESCRIPTION,
-        examples=["YTD"],
-    ),
-    benchmark_code: str | None = Query(
-        default=None,
-        description="Optional benchmark override used for relative concentration context.",
-        examples=["BMK_PB_GLOBAL_BALANCED_60_40"],
-    ),
-    as_of_date: str | None = Query(
-        default=None,
-        description="Optional business as-of date in YYYY-MM-DD format.",
-        examples=["2026-02-24"],
-    ),
-    report_start_date: str | None = Query(
-        default=None,
-        description=(
-            "Inclusive explicit start date when the caller requests an explicit risk window."
-        ),
-        examples=["2026-01-01"],
-    ),
-    report_end_date: str | None = Query(
-        default=None,
-        description="Inclusive explicit end date when the caller requests an explicit risk window.",
-        examples=["2026-03-27"],
-    ),
-    reporting_currency: str = Query(
-        default="USD",
-        description="Reporting currency used for stateful concentration analytics.",
-        examples=["USD"],
-    ),
-) -> WorkbenchRiskConcentrationResponse:
-    service = risk_workspace_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_concentration(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id,
-        period=period,
-        as_of_date=as_of_date,
-        report_start_date=report_start_date,
-        report_end_date=report_end_date,
-        reporting_currency=reporting_currency,
-        benchmark_code=benchmark_code,
     )
