+++ src/app/routers/workbench_risk_common.py
@@ -1,0 +1,5 @@
+RISK_PERIOD_QUERY_DESCRIPTION = (
+    "Canonical risk horizon. Use platform-governed values such as MTD, QTD, YTD, 1Y, 3Y, 5Y, "
+    "SI, YEAR, or EXPLICIT. Legacy aliases ONE_YEAR, THREE_YEAR, FIVE_YEAR, and ITD may be "
+    "accepted for compatibility but are normalized before calling lotus-risk."
+)
